requirements.txt
streamlit
python-dotenv
langchain
dotted_dict
faiss-cpu
langchain_community
langchain_openai
langchain_core 
langchain_text_splitters